+++ src/lambkin/common/exceptions.py
@@ -29,3 +29,32 @@ class LambkinSystemdNotFoundError(LambkinError):
 
 class LambkinSystemdScopeTimeoutError(LambkinError):
     """Raised when a systemd scope fails to stop within the timeout period."""
+
+
+class LambkinProcessDiedUnexpectedlyError(LambkinError):
+    """Raised when a background process exits before the context manager does.
+
+    Attributes:
+    ----------
+    argv : list[str]
+        The command that died.
+    returncode : int
+        The exit code of the process.
+    """
+
+    def __init__(self, argv: list[str], returncode: int) -> None:
+        """Initialize with the command and its exit code.
+
+        Parameters
+        ----------
+        argv : list[str]
+            The command that died.
+        returncode : int
+            The exit code of the process.
+        """
+        self.argv = argv
+        self.returncode = returncode
+        super().__init__(
+            f"Background process {argv[0]!r} died unexpectedly "
+            f"with return code {returncode}."
+        )
